requirements.txt
requests
pydantic
python-dotenv
pandas
sqlite-utils
beautifulsoup4
streamlit
altair
pyyaml
